Rework RunState and tick()
Reworks the RunState enum and how we move from state to state in our
GameState tick() function.

This is based on the final parts of Ch 7 / Part 2.6 from the tutorial.

This includes a workaround for an issue with the vision being rendered
out-of-date just after processing player input, causing a visual
"jitter". To fix this, we explicitly run the VisibilitySystem every time
after we call player_input(). I'd like to find a better fix for this.