Adds text console output
Extremely proof-of-concept demo of printing some text ("Hello!") to
screen by adding a new sparse console which uses the built-in "vga8x16.png"
font.

This does a bit of maths between the font size (8x16) and the tile
"font" size so the text is drawn at 1:1, instead of being scaled to the
tile size. This is needed because console width/height is specified in
characters/tiles instead of pixels.

I'd like to clean up this maths to avoid hardcoding values.

It may also help to define an enum for tracking console/layer IDs.