Adds UI w messaege log, reworks components
Adds "gui" module which abstracts what parts of the screen are used
for the camera view into the world and what is used for the new message
log view.

Adds the message log rendering with test lines. Reworks camera rendering
along with this.

Moves terminal builder to the new gui module.

Moves component registration and creating the player and monsters
to the components module. This reduces the main module's length and
dependencies. :)