Restructured gui and camera into ui module
Restructured the `gui` module into a new `ui` module, with functionality
split across multiple files for different parts of the interface.

The `gui.rs` file was much too big. It's now much easier to see what
imports are used for what, and to contain the rendering logic for
different parts of the UI.

`camera.rs` has also been moved into the `ui` module, as it had a lot of
the same concerns. It hasn't changed nearly as much, apart from updating
import paths and moving some constants to other files.