Revert "Install uy 0.1.2"
uy is maybe good for SI units, but it doesn't seem to have an easy
syntax for defining your own unit system... unless that unit system is
also a power-of-ten unit system like SI.

There is a different crate, `dimensioned`, which looks to be friendlier
and better suited to defining our own unit system. I'll try installing
that next.