Untis: test and fix Box2D.contains()
Our implementation was considering p2 and other positions on the "right"
and "bottom" edge as "contained" within the Box2D, but this isn't
correct with how we want to treat an integer-based Box2D.

Specifically, we don't want `for_each` to enumerate these positions, as
this would cause problems with map indexing in our game: a 10x10 map
should have positions ranging from 0,0 to 9,9, while being defined by
p1: 0,0 and p2: 10,10, with a width and height of 10,10. So for consistency,
a Box2D defined this way will not `contain()` p2 or any other positions
with those x/y coordinates.